files renamed in components
